Refactored I guess Created two model folders and moved files to it  Added TableCreator and TableTest
