+++ .gitignore
@@ -1,2 +1,4 @@
 Assignment1/.idea/*
 Assignment1/target/*
+.idea/*
+.idea/libraries/*
