+++ Assignment1/src/main/java/model/Card.java
@@ -1,4 +1,4 @@
-package controllers;
+package model;
 
 
 /**
