+++ Assignment1/src/test/java/model/DeckTest.java
@@ -1,4 +1,4 @@
-package controllers;
+package model;
 
 import static org.junit.Assert.*;
